Fixes for packet size, since we can do > 64b/pkt.

+++ libraries/UHS_host/UHS_MIDI/UHS_MIDI_INLINE.h
@@ -63,7 +63,7 @@ uint8_t UHS_NI UHS_MIDI::SetInterface(ENUMERATION_INFO *ei) {
                 if(ei->interface.epInfo[ep].bmAttributes == USB_TRANSFER_TYPE_BULK) {
                         uint8_t index = ((ei->interface.epInfo[ep].bEndpointAddress & USB_TRANSFER_DIRECTION_IN) == USB_TRANSFER_DIRECTION_IN) ? epDataInIndex : epDataOutIndex;
                         epInfo[index].epAddr = (ei->interface.epInfo[ep].bEndpointAddress & 0x0F);
-                        epInfo[index].maxPktSize = (uint8_t)(ei->interface.epInfo[ep].wMaxPacketSize);
+                        epInfo[index].maxPktSize = ei->interface.epInfo[ep].wMaxPacketSize;
                         epInfo[index].epAttribs = 0;
                         epInfo[index].bmNakPower = (index == epDataInIndex) ? UHS_USB_NAK_NOWAIT : UHS_USB_NAK_MAX_POWER;
                         epInfo[index].bmSndToggle = 0;
